+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.5] - 2025-11-12
+
+### Changed
+- Renamed gem from `acro_that` to `corp_pdf`
+- Renamed module from `AcroThat` to `CorpPdf`
+- Updated all internal references, documentation, and examples
+
 ## [1.0.4] - 2025-11-12
 
 ### Fixed
@@ -106,7 +113,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.1] - 2025-10-31
 
 ### Added
-- Initial release of AcroThat
+- Initial release of CorpPdf
 - Pure Ruby PDF AcroForm editing library
 - Support for parsing, listing, adding, removing, and modifying form fields
 - Signature field image appearance support (JPEG and PNG)